docs: README.md 업데이트 및 소수점/페니 주식 매매 지원을 위한 Double 전환 및 안정성 향상 패치 적용

+++ account_server_kt/src/main/kotlin/com/mts/account/service/LedgerService.kt
@@ -100,7 +100,7 @@ class LedgerService(
             buyerId = buyerId,
             sellerId = sellerId,
             ticker = ticker,
-            price = price.toInt(),
+            price = price,
             quantity = quantity
         ))
     }
